+++ rewatch/tests/compile.sh
@@ -52,51 +52,51 @@ popd > /dev/null
 node ./packages/main/src/Main.mjs > ./packages/main/src/output.txt
 
 mv ./packages/main/src/Main.res ./packages/main/src/Main2.res
-rewatch build --snapshot-output &> ../tests/snapshots/rename-file.txt
+rewatch build &> ../tests/snapshots/rename-file.txt
 normalize_paths ../tests/snapshots/rename-file.txt
 mv ./packages/main/src/Main2.res ./packages/main/src/Main.res
 
 # Rename a file with a dependent - this should trigger an error
 mv ./packages/main/src/InternalDep.res ./packages/main/src/InternalDep2.res
-rewatch build --snapshot-output &> ../tests/snapshots/rename-file-internal-dep.txt
+rewatch build &> ../tests/snapshots/rename-file-internal-dep.txt
 normalize_paths ../tests/snapshots/rename-file-internal-dep.txt
 mv ./packages/main/src/InternalDep2.res ./packages/main/src/InternalDep.res
 
 # Rename a file with a dependent in a namespaced package - this should trigger an error (regression)
 mv ./packages/new-namespace/src/Other_module.res ./packages/new-namespace/src/Other_module2.res
-rewatch build --snapshot-output &> ../tests/snapshots/rename-file-internal-dep-namespace.txt
+rewatch build &> ../tests/snapshots/rename-file-internal-dep-namespace.txt
 normalize_paths ../tests/snapshots/rename-file-internal-dep-namespace.txt
 mv ./packages/new-namespace/src/Other_module2.res ./packages/new-namespace/src/Other_module.res
 
 rewatch build &>  /dev/null
 mv ./packages/main/src/ModuleWithInterface.resi ./packages/main/src/ModuleWithInterface2.resi
-rewatch build --snapshot-output &> ../tests/snapshots/rename-interface-file.txt
+rewatch build &> ../tests/snapshots/rename-interface-file.txt
 normalize_paths ../tests/snapshots/rename-interface-file.txt
 mv ./packages/main/src/ModuleWithInterface2.resi ./packages/main/src/ModuleWithInterface.resi
 rewatch build &> /dev/null
 mv ./packages/main/src/ModuleWithInterface.res ./packages/main/src/ModuleWithInterface2.res
-rewatch build --snapshot-output &> ../tests/snapshots/rename-file-with-interface.txt
+rewatch build &> ../tests/snapshots/rename-file-with-interface.txt
 normalize_paths ../tests/snapshots/rename-file-with-interface.txt
 mv ./packages/main/src/ModuleWithInterface2.res ./packages/main/src/ModuleWithInterface.res
 rewatch build &> /dev/null
 
 # when deleting a file that other files depend on, the compile should fail
 rm packages/dep02/src/Dep02.res
-rewatch build --snapshot-output &> ../tests/snapshots/remove-file.txt
+rewatch build &> ../tests/snapshots/remove-file.txt
 normalize_paths ../tests/snapshots/remove-file.txt
 git checkout -- packages/dep02/src/Dep02.res
 rewatch build &> /dev/null
 
 # it should show an error when we have a dependency cycle
 echo 'Dep01.log()' >> packages/new-namespace/src/NS_alias.res
-rewatch build --snapshot-output &> ../tests/snapshots/dependency-cycle.txt
+rewatch build &> ../tests/snapshots/dependency-cycle.txt
 normalize_paths ../tests/snapshots/dependency-cycle.txt
 git checkout -- packages/new-namespace/src/NS_alias.res
 
 # this should not compile because  "@rescript/webapi" is part of dev-dependencies
 # and FileToTest.res is not listed as "type":"dev"
 echo 'open WebAPI' >> packages/with-dev-deps/src/FileToTest.res
-rewatch build --snapshot-output &> ../tests/snapshots/bs-dev-dependency-used-by-non-dev-source.txt
+rewatch build &> ../tests/snapshots/bs-dev-dependency-used-by-non-dev-source.txt
 normalize_paths ../tests/snapshots/bs-dev-dependency-used-by-non-dev-source.txt
 git checkout -- packages/with-dev-deps/src/FileToTest.res